requirements.txt
flask==2.3.5
python-dotenv==1.0.1
google-generativeai==0.1.5
requests==2.31.0
